pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "proactive-affective-agent"
version = "0.1.0"
description = "Proactive Affective Agent: passive sensing → emotional state prediction → just-in-time intervention"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "pandas>=2.0",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "faiss-cpu>=1.7",
    "openai>=1.0",
    "anthropic>=0.20",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "tqdm>=4.65",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "jupyter>=1.0",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "ipykernel>=6.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
